loan_tests.py
import unittest
import numpy as np
from unittest.mock import patch, mock_open
import loan  # Import the neural network module you've written
import pandas as pd

class TestNeuralNetwork(unittest.TestCase):
    """
        This is simply some examples of testing you could do.
    """

    def setUp(self):
        # Set a seed for reproducibility
        np.random.seed(42)
        #self.mock_data = "clientid,income,age,loan,known_result\n1,50000,25,10000,0\n2,60000,35,15000,1"

    def test_sigmoid_output_range(self):
        # Sigmoid should always output between 0 and 1
        for x in np.linspace(-10, 10, 100):
            result = loan.sigmoid(x)
            self.assertTrue(0 <= result <= 1)

    def test_sigmoid_derivative(self):
        # Sigmoid derivative tested on known values
        values = np.array([0.2, 0.5, 0.8])
        expected = values * (1 - values)
        np.testing.assert_array_almost_equal(loan.sigmoid_derivative(values), expected)

    def test_calculate_output_shape(self):
        # Assuming weights have been initialized with controlled randomness
     loan.weights0 = 2 * np.random.random((3, 13)) - 1
     loan.weights1 = 2 * np.random.random((13, 1)) - 1
     instance = np.array([0.5, 0.3, 0.2])
     output = loan.calculate_output(instance, loan.weights0, loan.weights1)
     self.assertEqual(isinstance(output, float), True)

    @patch('pandas.read_csv')
    def test_input_handling(self, mock_read_csv):
        # Mock reading CSV and check how NaNs are handled
        mock_read_csv.return_value = pd.DataFrame({
            'i#clientid': [1,2,3],
            'income': [50000, 60000, None],
            'age': [25, 5, 35],
            'loan': [None, 15000, 20000],
            'c#default': [0,0,0]
        })
        inputs, outputs = loan.prepare_data('dummy_path')
        self.assertEqual(len(inputs), 1)  # Should drop row with NaN

# Additional tests can include testing learning parameters influence, error handling for corrupted data, etc.

if __name__ == '__main__':
    unittest.main()
